% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{add_charts}
\alias{add_charts}
\title{Create R Markdown code for monthly option day trade time series}
\usage{
add_charts(ac, cap)
}
\arguments{
\item{ac}{is account number}

\item{cap}{is the value of portfolio}
}
\value{
text of R Markdown code
}
\description{
Create R Markdown code for monthly option day trade time series
}
