% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{add_charts_month}
\alias{add_charts_month}
\title{Create R Markdown code for daily charts for specific month}
\usage{
add_charts_month(m, value)
}
\arguments{
\item{m}{is the month number (1-12)}

\item{value}{is the value of portfolio}
}
\value{
text of R Markdown code
}
\description{
Create R Markdown code for daily charts for specific month
}
