% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{add_charts_year}
\alias{add_charts_year}
\title{Create R Markdown code for monthly charts for specific year}
\usage{
add_charts_year(y, value)
}
\arguments{
\item{y}{is the year number (2024)}

\item{value}{is the value of portfolio}
}
\value{
text of R Markdown code
}
\description{
Create R Markdown code for monthly charts for specific year
}
