% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{add_futures_table_year}
\alias{add_futures_table_year}
\title{Create R Markdown code for futures ticker statistics for specific year as a table}
\usage{
add_futures_table_year()
}
\value{
text of R Markdown code
}
\description{
Create R Markdown code for futures ticker statistics for specific year as a table
}
