% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{add_general_statistics}
\alias{add_general_statistics}
\title{Add R Markdown code for general statistics}
\usage{
add_general_statistics(cap, gStat)
}
\arguments{
\item{cap}{is the initial value}

\item{gStat}{is the vector that has the following data:
plTotal,commisTotal,yield,pcYield,nTrades,winners,pcWinners,contracts,
contrTrade,maxWin,maxLoss,avgCapRisk,avgRoi,
avgDuration,maxDuration,minDuration,trDays,trDaysUsed,pctrDaysUsed}
}
\value{
text of R Markdown code
}
\description{
Add R Markdown code for general statistics
}
