% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{add_header}
\alias{add_header}
\title{Create an R Markdown code for the document header}
\usage{
add_header(title, author, date)
}
\arguments{
\item{title}{is the title of the R Markdown file}

\item{author}{is the author of R Markdown file}

\item{date}{is the date of release}
}
\value{
The text for R Markdown title (YAML fragment)
}
\description{
Create an R Markdown code for the document header
}
