% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{add_prefix3}
\alias{add_prefix3}
\title{Include an R code into R Markdown file}
\usage{
add_prefix3(cap, tradesFile)
}
\arguments{
\item{cap}{is initial value of portfolio (capital)}

\item{tradesFile}{is the name of csv file with option day trade data}
}
\value{
text of R Markdown code
}
\description{
Include an R code into R Markdown file
}
