% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{add_table}
\alias{add_table}
\title{Create R Markdown code for ticker statistics as a table}
\usage{
add_table()
}
\value{
text of R Markdown code
}
\description{
Create R Markdown code for ticker statistics as a table
}
