% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{add_table_month}
\alias{add_table_month}
\title{Create R Markdown code for ticker statistics for specific month as a table}
\usage{
add_table_month()
}
\value{
text of R Markdown code
}
\description{
Create R Markdown code for ticker statistics for specific month as a table
}
