% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{add_title_n}
\alias{add_title_n}
\title{Add R Markdown code for n-level header}
\usage{
add_title_n(text, n, tabset = FALSE)
}
\arguments{
\item{text}{is the header's name}

\item{n}{is the header's level (1-5)}

\item{tabset}{if =TRUE adds tabset parameters to the header}
}
\value{
text of R Markdown code
}
\description{
Add R Markdown code for n-level header
}
