% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{checkdate}
\alias{checkdate}
\title{Checks if a string a valid date}
\usage{
checkdate(dd)
}
\arguments{
\item{dd}{is a string of format mm-dd-yyyy}
}
\value{
a logical value TRUE or FALSE
}
\description{
Checks if a string a valid date
}
