% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{countCallsPuts}
\alias{countCallsPuts}
\title{Counts option trades for an underlying symbol}
\usage{
countCallsPuts(TRADES, symb)
}
\arguments{
\item{TRADES}{The data frame of trades with the following fields:
"accountId","entryDate","exitDate","underSymbol","symbol",
"entryPrice","exitPrice","amount","entryCommis","exitCommis","pl",
"winner","capAtRisk","roi","tradeDur","year","month","day"}

\item{symb}{is an underlying symbol}
}
\value{
The data frame with the following fields:
"symb","nTrades","nWins","pcWins","yield",
"nCalls","nCallsWins","pcCallsWins",
"nPuts","nPutsWins","pcPutsWins"
}
\description{
Counts option trades for an underlying symbol
}
