% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{countLongsShorts}
\alias{countLongsShorts}
\title{Counts futures trades for a future symbol}
\usage{
countLongsShorts(TRADES, symb)
}
\arguments{
\item{TRADES}{The data frame of trades with the following fields:
"accountId","entryDate","exitDate","underSymbol","symbol",
"entryPrice","exitPrice","amount","entryCommis","exitCommis","pl",
"winner","capAtRisk","roi","tradeDur","year","month","day"}

\item{symb}{is a future symbol}
}
\value{
The data frame with the following fields:
"symb","nTrades","nWins","pcWins","yield",
"nLong","nLongWins","pcLongWins",
"nShort","nShortWins","pcShortWins"
}
\description{
Counts futures trades for a future symbol
}
