% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading-package.R
\docType{package}
\name{daytrading-package}
\alias{daytrading}
\alias{daytrading-package}
\title{daytrading: Functions for creating day trading report}
\description{
This package has functions for generating day trading report.
}
\author{
\strong{Maintainer}: Valery Petrushin \email{vapetr3@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
