% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{generateFuturesAccountCode}
\alias{generateFuturesAccountCode}
\title{Generate R Markdown code for an account's futures day trading}
\usage{
generateFuturesAccountCode(
  tradesFile,
  TRADES,
  ac,
  curDate,
  WD,
  cap,
  startDate,
  endDate,
  isTest = FALSE,
  title = "Futures Day Trading Report"
)
}
\arguments{
\item{tradesFile}{is the file name of one day futures trades}

\item{TRADES}{is the frame of trade data with the following fields:
"accountId","entryDate","exitDate","underSymbol","symbol","entryPrice","exitPrice","amount",
"entryCommis","exitCommis","pl","winner","capAtRisk","roi","tradeDur","year","month","day"}

\item{ac}{is the account number}

\item{curDate}{is the current date (format yyyy-mm-dd)}

\item{WD}{is the frame of market working dates with the following fields:
"dates","dayOfWeek","workDates"}

\item{cap}{is the value of account at the beginning}

\item{startDate}{is the first global trade date}

\item{endDate}{is the last global trade date}

\item{isTest}{prints additional info if TRUE (default is FALSE)}

\item{title}{is the title of day trading report (default is "Futures Day Trading Report")}
}
\value{
a string of R Markdown code for the account's trades
}
\description{
Generate R Markdown code for an account's futures day trading
}
