% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{generateMonthCode}
\alias{generateMonthCode}
\title{Generate R Markdown code for monthly option day trading}
\usage{
generateMonthCode(TRADES, m, WD, value, startDate, endDate, isTest = FALSE)
}
\arguments{
\item{TRADES}{is the frame of trade data with the following fields:
"accountId","entryDate","exitDate","underSymbol","symbol","entryPrice","exitPrice","amount",
"entryCommis","exitCommis","pl","winner","capAtRisk","roi","tradeDur","year","month","day"}

\item{m}{is the number of the month (1-12)}

\item{WD}{is the frame of market working dates with the following fields:
"dates","dayOfWeek","workDates"}

\item{value}{is the value of portfolio at the beginning of the month}

\item{startDate}{is the first global trade date}

\item{endDate}{is the last global trade date}

\item{isTest}{prints additional info if TRUE (default is FALSE)}
}
\value{
a structure with two fields: $txt is R Markdown code and $value is portfolio value at the end of month
}
\description{
Generate R Markdown code for monthly option day trading
}
