% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{generateOneDayReport}
\alias{generateOneDayReport}
\title{Generates low level report for option day trades as a csv file}
\usage{
generateOneDayReport(TRADES, header = TRUE)
}
\arguments{
\item{TRADES}{The data frame of trades with the following fields:
"accountId","entryDate","exitDate","underSymbol","symbol",
"entryPrice","exitPrice","amount","entryCommis","exitCommis","pl",
"winner","capAtRisk","roi","tradeDur","year","month","day"}

\item{header}{if = TRUE outputs the header (default is TRUE)}
}
\value{
The data frame with the following columns:
"Action","Trade Date","Stock","","Option","Quantity","Price","Stop",
"Target","Gross Entry/Exit","# contracts","Commission","Gross P or L",
"Winner","Loser","Capital At Risk","Porfolio Size","","ROI","Win \%",
"Loss \%","Duration Of Signal","Win $","Loss $"
}
\description{
Processes trades to get option day report for each trade
}
