% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{generateYearCode}
\alias{generateYearCode}
\title{Generate R Markdown code for yearly option day trading}
\usage{
generateYearCode(
  TRADES,
  y,
  WD,
  value,
  startDate,
  endDate,
  curDate,
  isTest = FALSE
)
}
\arguments{
\item{TRADES}{is the frame of trade data with the following fields:
"accountId","entryDate","exitDate","underSymbol","symbol","entryPrice","exitPrice","amount",
"entryCommis","exitCommis","pl","winner","capAtRisk","roi","tradeDur","year","month","day"}

\item{y}{is the number of the year (example: 2024)}

\item{WD}{is the frame of market working dates with the following fields:
"dates","dayOfWeek","workDates"}

\item{value}{is the value of portfolio at the beginning of the year}

\item{startDate}{is the first global trade date}

\item{endDate}{is the last global trade date}

\item{curDate}{is thr current date}

\item{isTest}{prints additional info if TRUE (default is FALSE)}
}
\value{
a string of R Markdown code for the yearly trades
}
\description{
Generate R Markdown code for yearly option day trading
}
