% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{getLastName}
\alias{getLastName}
\title{Gets the last name}
\usage{
getLastName(name)
}
\arguments{
\item{name}{is a string that has a name separated by "."}
}
\value{
the substring after the last "."
}
\description{
Gets the last name
}
\examples{
getLastName("aaa.bbb.ccc")  # returns "ccc"

}
