% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{getTickerStatistics}
\alias{getTickerStatistics}
\title{Get statistics for all option trades for underlying symbol}
\usage{
getTickerStatistics(TRADES)
}
\arguments{
\item{TRADES}{The data frame of trades with the following fields:
"accountId","entryDate","exitDate","underSymbol","symbol",
"entryPrice","exitPrice","amount","entryCommis","exitCommis","pl",
"winner","capAtRisk","roi","tradeDur","year","month","day"}
}
\value{
The data frame with the following fields:
'symbol','nTrades','nWins','pcWins','yield',
'nCalls','nCallsWins','pcCallsWins',
'nPuts','nPutsWins','pcPutsWins'
}
\description{
Get statistics for all option trades for underlying symbol
}
