% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{getTimeSeries}
\alias{getTimeSeries}
\title{Create trade time series data frame on monthly or daily level}
\usage{
getTimeSeries(TRADES, WD, cap, startDate, endDate, level = "total")
}
\arguments{
\item{TRADES}{The data frame of trades with the following fields:
"accountId","entryDate","exitDate","underSymbol","symbol",
"entryPrice","exitPrice","amount","entryCommis","exitCommis","pl",
"winner","capAtRisk","roi","tradeDur","year","month","day"}

\item{WD}{is the frame of market working dates with the following fields:
"dates","dayOfWeek","workDates"}

\item{cap}{is the portfolio value (capital)}

\item{startDate}{is the first global trade date}

\item{endDate}{is the last global trade date}

\item{level}{if = 'total' then data is aggregated to monthly otherwise to daily}
}
\value{
the time series data frame with the following fields:
'time','value','yield','ntrades','winners','pcwinners',
'contrtrade','roi','duration'
}
\description{
Create trade time series data frame on monthly or daily level
}
