% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{getTradeDate}
\alias{getTradeDate}
\title{Splits date/time data}
\usage{
getTradeDate(dd)
}
\arguments{
\item{dd}{The date like  "2023-10-19 14:32:30"}
}
\value{
The date part ("2023-10-19")
}
\description{
Gets the date part of the date
}
