% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{getTradeDateMonth}
\alias{getTradeDateMonth}
\title{Splits date data}
\usage{
getTradeDateMonth(dd)
}
\arguments{
\item{dd}{The date like  "2023-10-19"}
}
\value{
The month part ("10")
}
\description{
Gets the month part of the date
}
