% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{getTradeDateYear}
\alias{getTradeDateYear}
\title{Splits date data}
\usage{
getTradeDateYear(dd)
}
\arguments{
\item{dd}{The date like  "2023-10-19"}
}
\value{
The year part ("2023")
}
\description{
Gets the year part of the date
}
