% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{getTransactions}
\alias{getTransactions}
\title{Gets trade transaction for the specified account}
\usage{
getTransactions(tokens, accHash, startDate, endDate, assetType = "OPTION")
}
\arguments{
\item{tokens}{are access and refresh tokens for account}

\item{accHash}{is a hash value for specific account number}

\item{startDate}{is the start date for transactions}

\item{endDate}{is the end date for transactions}

\item{assetType}{is the assete type ('OPTION' (default),'FUTURE' or 'EQUITY')}
}
\value{
a data frame with transactions
}
\description{
Gets trade transaction for the specified account
}
