% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{processFuturesOneDayTrades}
\alias{processFuturesOneDayTrades}
\title{Processes transactions to get futures day trades}
\usage{
processFuturesOneDayTrades(TRANS)
}
\arguments{
\item{TRANS}{The data frame of transactions with the following fields:
activityId,time,accountNumber,type,status,subAccount,tradeDate,orderId,
netAmount,transferItems,assetType,status,symbol,description,instrumentId,
closingPrice,activeContract,expirationDate,lastTradingDate,multiplier,
futureType,amount,cost,price,positionEffect}
}
\value{
The data frame with the following columns:
"accountId","entryDate","exitDate","underSymbol","symbol",
"entryPrice","exitPrice","amount","entryCommis","exitCommis","pl",
"winner","capAtRisk","roi","tradeDur","year","month","day"
}
\description{
Processes transactions to get futures day trades
}
