% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{processOneDayTrades}
\alias{processOneDayTrades}
\title{Gets day trades}
\usage{
processOneDayTrades(TRANS)
}
\arguments{
\item{TRANS}{The data frame of transactions with the following fields:
,index,activityId,time,accountNumber,type,status,subAccount,tradeDate,orderId,
netAmount,positionId,transferItems.amount,transferItems.cost,transferItems.price,
transferItems.instrument.assetType,transferItems.instrument.status,transferItems.
instrument.symbol,transferItems.instrument.description,transferItems.instrument.instrumentId,
transferItems.instrument.expirationDate,transferItems.instrument.optionDeliverables,
transferItems.instrument.optionPremiumMultiplier,transferItems.instrument.putCall,
transferItems.instrument.strikePrice,transferItems.instrument.type,transferItems.
instrument.underlyingSymbol,transferItems.instrument.underlyingCusip,transferItems.feeType,
transferItems.instrument.closingPrice,transferItems.instrument.activeContract,
transferItems.instrument.lastTradingDate,transferItems.instrument.multiplier,
transferItems.instrument.futureType,transferItems.positionEffect}
}
\value{
The data frame with the following columns:
"accountId","entryDate","exitDate","underSymbol","symbol",
"entryPrice","exitPrice","amount","entryCommis","exitCommis","pl",
"winner","capAtRisk","roi","tradeDur","year","month","day"
}
\description{
Processes transactions to get option day trades
}
