% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{select_accountNumbers}
\alias{select_accountNumbers}
\title{Analyze and select account numbers}
\usage{
select_accountNumbers(accNums, PARS)
}
\arguments{
\item{accNums}{a data frame of Schwab account(s)
with the fields 'accountNumber' and 'hashValue'}

\item{PARS}{a data frame from parameter file
.       with the fields 'account.number','account.initValue' and 'account.startDate'}
}
\value{
a data frame with the following fields
'accountNumber','initValue','startDate','hashValue'
}
\description{
Analyze and select account numbers
}
