% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{splitTransDate}
\alias{splitTransDate}
\title{Splits Schwab's transaction date and gets the date part of it}
\usage{
splitTransDate(dd)
}
\arguments{
\item{dd}{The date like  "2023-08-07T18:15:32+0000"}
}
\value{
The date part ("2023-08-07")
}
\description{
Splits Schwab's transaction date and gets the date part of it
}
