% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{splitTransTime}
\alias{splitTransTime}
\title{Splits Schwab's transaction date and gets the time part of it}
\usage{
splitTransTime(dd)
}
\arguments{
\item{dd}{The date like  "2023-08-07T18:15:32+0000"}
}
\value{
The time part ("18:15:32+0000")
}
\description{
Splits Schwab's transaction date and gets the time part of it
}
