% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{toUSD}
\alias{toUSD}
\title{Converts a number to US dollar format}
\usage{
toUSD(money)
}
\arguments{
\item{money}{a number}
}
\value{
a string that represent number as a dollar value
}
\description{
Converts a number to US dollar format
}
\examples{
toUSD(109245.65)  # "$109,245.65"
toUSD(-10924.6)   # "-$10,924.60"
}
