% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{getParams}
\alias{getParams}
\title{Read the params.yml file that specifies parameters}
\usage{
getParams()
}
\value{
The data frame with the following fields:
"account.number", "account.initValue", "account.startDate"
}
\description{
\subsection{The YAML file params.yml may have parameters for one or more accounts.
The multiple accounts should be linked.
It has the following structure:}{
\itemize{
\item account:
number: '99999999'
initValue: 25000
startDate: 2024-01-01
. . .
\item account:
number: '88888888'
initValue: 30000
startDate: 2024-01-01
}
}
}
