% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{getTradingDates}
\alias{getTradingDates}
\title{Create the data frame of market working dates for 2018-2025}
\usage{
getTradingDates()
}
\value{
The data frame that has the folloeing fields:
dates is date in the formay 'yyy-mm-dd'
dayOfWeek is the corresponding day of week Sunday - Saturday
workDates = 1 if it is market work day or = 0 otherwise
}
\description{
Create the data frame of market working dates for 2018-2025
}
\examples{
      WD = getTradingDates()
}
