% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytrading.R
\name{getTradingStatistics}
\alias{getTradingStatistics}
\title{Calculate statistics for option daily trades on all levels
total - monthly - daily}
\usage{
getTradingStatistics(TRADES, WD, cap)
}
\arguments{
\item{TRADES}{The data frame of trades with the following fields:
"accountId","entryDate","exitDate","underSymbol","symbol",
"entryPrice","exitPrice","amount","entryCommis","exitCommis","pl",
"winner","capAtRisk","roi","tradeDur","year","month","day"}

\item{WD}{is the frame of market working dates with the following fields:
"dates","dayOfWeek","workDates"}

\item{cap}{is the portfolio value (capital)}
}
\value{
The statistics data frame with the following fields:
stats = c(plTotal,commisTotal,yield,pcYield,nTrades,winners,pcWinners,
contracts,contrTrade,maxWin,maxLoss,avgCapRisk,avgRoi,
avgDuration,maxDuration,minDuration,trDays,trDaysUsed,pctrDaysUsed
}
\description{
Calculate statistics for option daily trades on all levels
total - monthly - daily
}
